deception/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
streamlit==1.40.0
transformers==4.45.2
torch==2.7.1
python-multipart==0.0.12
aiofiles==24.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
plotly==5.24.1
discord-webhook==1.3.1
aiosmtplib==3.0.2
jinja2==3.1.4
folium==0.17.0
streamlit-folium==0.22.0
numpy==2.2.1
ipaddress==1.0.23